// src/modules/notifications/enums/notification-type.enum.ts
export enum NotificationType {
  PURCHASE_REQUEST_APPROVAL = 'PURCHASE_REQUEST_APPROVAL',
  PURCHASE_REQUEST_STATUS = 'PURCHASE_REQUEST_STATUS',
  PURCHASE_REQUEST_PURCHASING = 'PURCHASE_REQUEST_PURCHASING',
  PURCHASE_REQUEST_PURCHASED = 'PURCHASE_REQUEST_PURCHASED',
  WAREHOUSE_RECEIPT = 'WAREHOUSE_RECEIPT',
  TRANSFER_RECEIVED = 'TRANSFER_RECEIVED',
  PERMISSIONS_GRANTED = 'PERMISSIONS_GRANTED',
}